cosmetic
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@11393 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/Application.cpp
@@ -4843,7 +4843,7 @@ void CApplication::Restart(bool bSamePosition)
   CStdString state = m_pPlayer->GetPlayerState();
 
   // set the requested starttime
-  m_itemCurrentFile.m_lStartOffset = time * 75.0;
+  m_itemCurrentFile.m_lStartOffset = (long)(time * 75.0);
 
   // reopen the file
   if ( PlayFile(m_itemCurrentFile, true) && m_pPlayer )
